bioinfo-cnn/requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
scikit-learn==1.3.2
pillow==10.2.0
tqdm==4.66.2
captum==0.7.0  # For model interpretability
albumentations==1.3.1  # Downgraded for compatibility
albucore==0.0.13  # Specific version for compatibility 